moved things around in the examples directory; updates #138'
