+++ examples/swmr/swmr_builder.cpp
