+++ examples/swmr/swmr_environment.cpp
