+++ examples/swmr/swmr_reader.cpp
