+++ examples/swmr/swmr_write.cpp
